+++ CHANGELOG.md
@@ -21,6 +21,13 @@ nodes in AzureML. Example: Add `--num_nodes=2` to the commandline arguments to t
   if `CHANGELOG.md` has been modified.
 - ([#412](https://github.com/microsoft/InnerEye-DeepLearning/pull/412)) Dataset files can now have arbitrary names, and are no longer restricted to be called
   `dataset.csv`, via the config field `dataset_csv`. This allows to have a single set of image files in a folder, but multiple datasets derived from it. 
+- ([#391](https://github.com/microsoft/InnerEye-DeepLearning/pull/391)) Support for multilabel classification tasks. 
+  Multilabel models can be trained by adding the parameter `class_names` to the config for classification models. 
+  `class_names` should contain the name of each label class in the dataset, and the order of names should match the 
+  order of class label indices in `dataset.csv`.
+  `dataset.csv` supports multiple labels (indices corresponding to `class_names`) per subject in the label column. 
+  Multiple labels should be encoded as a string with labels separated by a `|`, for example "0|2|4".
+  Note that this PR does not add support for multiclass models, where the labels are mutually exclusive.
 
 ### Changed
 - ([#385](https://github.com/microsoft/InnerEye-DeepLearning/pull/385)) Starting an AzureML run now uses the